[ising] format the sycl program slightly; call the single-precision exponent function in the kernel

+++ ising-cuda/main.cu
@@ -81,7 +81,7 @@ __global__ void update_lattice(signed char* lattice,
 
   // Determine whether to flip spin
   signed char lij = lattice[i * ny + j];
-  float acceptance_ratio = exp(-2.0f * inv_temp * nn_sum * lij);
+  float acceptance_ratio = expf(-2.0f * inv_temp * nn_sum * lij);
   if (randvals[i*ny + j] < acceptance_ratio) {
     lattice[i * ny + j] = -lij;
   }
@@ -203,7 +203,6 @@ int main(int argc, char **argv) {
   }
 
   float inv_temp = 1.0f / (alpha*TCRIT);
-
 
 #ifdef CURAND
   // Setup cuRAND generator
